requirements.txt
Flask==2.1.1
psycopg2-binary>=2.9.9
werkzeug==2.3.8
bs4
PyYAML
flask-cors
flask-swagger-ui
python-dateutil
